@@ -24,19 +24,18 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_breadcrumb: ^0.1.2
+  flutter_form_builder: ^4.1.0
   flutter_markdown: ^0.4.3
   flutter_secure_storage: ^3.3.3
   flutter_signin_button: ^1.0.0
   flutter_svg: ^0.18.0
-  google_sign_in: ^4.5.3
-  graphql_flutter: ^3.0.0
-  photo_view: ^0.10.0
+  dropdown_search: ^0.4.8
+  google_sign_in: ^4.5.6
+  graphql_flutter: ^3.1.0
+  photo_view: ^0.10.3
   provider: ^4.3.2
   shared_preferences: ^0.5.8
   sliding_up_panel: ^1.0.2
-  # sliding_up_panel:
-  #   git:
-  #     url: git://github.com/Adem68/sliding_up_panel.git
 
 
   # The following adds the Cupertino Icons font to your application.
