@@ -2,24 +2,22 @@ import 'package:flutter/material.dart';
 import 'package:sliding_up_panel/sliding_up_panel.dart';
 
 class PreviewPanel extends StatelessWidget {
-  PreviewPanel(
-      {@required this.main,
-      @required this.preview,
-      this.title = 'preview',
-      this.controller});
+  PreviewPanel({@required this.main, @required this.preview, this.controller});
   final Widget main;
   final Widget preview;
-  final String title;
   final PanelController controller;
   @override
   Widget build(BuildContext context) {
+    final color = Theme.of(context).colorScheme.secondary;
     return SlidingUpPanel(
         controller: controller,
-        minHeight: 45,
+        margin: EdgeInsets.fromLTRB(15, 0, 15, 0),
+        minHeight: 35,
         backdropEnabled: true,
         color: Theme.of(context).cardColor,
         borderRadius: BorderRadius.only(
-            topLeft: Radius.circular(20), topRight: Radius.circular(20)),
+            topLeft: Radius.circular(10), topRight: Radius.circular(10)),
+        border: Border.all(color: color.withOpacity(0.65)),
         body: main,
         panel: Container(
             padding: EdgeInsets.all(10),
@@ -29,10 +27,7 @@ class PreviewPanel extends StatelessWidget {
                     height: 3.0,
                     width: 80,
                     color: Theme.of(context).colorScheme.secondary),
-                SizedBox(height: 3),
-                Text(title,
-                    style: TextStyle(color: Colors.white, fontSize: 18.0)),
-                SizedBox(height: 8),
+                SizedBox(height: 20),
                 preview,
               ],
             )));
