@@ -1,11 +1,18 @@
+use crate::models::PlotType;
 use async_graphql::SimpleObject;
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 use uuid::Uuid;
 
-pub type Db = sqlx::PgPool;
 pub type Json = serde_json::Value;
 pub type GenericError = Box<dyn std::error::Error + Send + Sync + 'static>;
+pub type Vars<'v> = [(&'v str, &'v str)];
+
+#[derive(Debug, Clone)]
+pub struct Db {
+    pub meta: sqlx::PgPool,
+    pub data: sqlx::PgPool,
+}
 
 pub struct ModelKeys {
     pub model: String,
@@ -16,6 +23,15 @@ pub struct ModelKeys {
 pub struct UploadDatasetPayload {
     pub uri: String,
     pub uuid: Uuid,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct CreatePlotPayload {
+    pub view: String,
+    pub uuid: Uuid,
+    #[serde(rename = "type")]
+    pub type_: PlotType,
+    pub args: Json,
 }
 
 #[derive(
