@@ -7,23 +7,25 @@ import 'nav.dart';
 
 class Project extends StatelessWidget {
   final List<AdaptiveButtonData> buttonData;
-  Project({Nav nav, String projectID})
+  Project({Nav nav, String projectId})
       : buttonData = [
           AdaptiveButtonData(
               icon: Icons.storage,
               name: 'datasets',
               onTap: () => nav.push('datasets',
-                  d.Datasets(projectID: projectID), d.makeAdder(projectID))),
+                  d.Datasets(projectId: projectId), d.makeAdder(projectId))),
           AdaptiveButtonData(
               icon: Icons.assessment,
               name: 'analyses',
-              onTap: () => nav.push('analyses',
-                  a.Analyses(nav: nav, projectID: projectID), a.add)),
+              onTap: () => nav.push(
+                  'analyses',
+                  a.Analyses(nav: nav, projectId: projectId),
+                  a.makeAdder(projectId))),
           AdaptiveButtonData(
               icon: Icons.lock,
               name: 'permissions',
               onTap: () => nav.push(
-                  'permissions', p.Permissions(projectID: projectID), p.add)),
+                  'permissions', p.Permissions(projectId: projectId), p.add)),
         ];
   @override
   Widget build(BuildContext context) {
